+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+
+## [Unreleased]
+#### Fixed
+- Details in createChangeRoleRequest at unblock user
+
 ## [1.12.0] - 2020-03-17
 #### Added
 - Ability change trailing digits in asset