+++ CHANGELOG.md
@@ -9,6 +9,7 @@ for further information about branching and tagging conventions.
 
 ## [Unreleased]
 #### Fixed
+- A bug when non-withdrawable asset's input fields are active in withdrawal form
 - Details in createChangeRoleRequest at unblock user
 
 ## [1.12.0] - 2020-03-17