+++ CHANGELOG.md
@@ -10,6 +10,7 @@ for further information about branching and tagging conventions.
 ## [Unreleased]
 #### Fixed
 - Details in createChangeRoleRequest at unblock user
+- Disable submit button until Document type is invalid
 
 ## [1.12.0] - 2020-03-17
 #### Added