+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
-## [Unreleased]
+## [1.12.1-rc.0] - 2020-03-25
 #### Fixed
 - Details in createChangeRoleRequest at unblock user
 - A bug with localization
@@ -549,7 +549,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/admin-panel/compare/1.12.0...HEAD
+[Unreleased]: https://github.com/tokend/admin-panel/compare/1.12.1-rc.0...HEAD
+[1.12.1-rc.0]: https://github.com/tokend/admin-panel/compare/1.12.0...1.12.1-rc.0
 [1.12.0]: https://github.com/tokend/admin-panel/compare/1.12.0-rc.4...1.12.0
 [1.12.0-rc.4]: https://github.com/tokend/admin-panel/compare/1.12.0-rc.3...1.12.0-rc.4
 [1.12.0-rc.3]: https://github.com/tokend/admin-panel/compare/1.12.0-rc.2...1.12.0-rc.3